raise FileExistsError if not replace==true

+++ server/frameioclient/download.py
@@ -1,11 +1,11 @@
 import requests
-import math
 import os
 
 class FrameioDownloader(object):
-  def __init__(self, asset, download_folder):
+  def __init__(self, asset, download_folder, replace):
     self.asset = asset
     self.download_folder = download_folder
+    self.replace = replace
 
   def download(self):
     original_filename = self.asset['name']
@@ -13,6 +13,9 @@ class FrameioDownloader(object):
     
     url = self.asset['original']
     r = requests.get(url)
-    
+
+    if os.path.isfile(final_destination) and not self.replace:
+        raise FileExistsError
+
     open(final_destination, 'wb').write(r.content)
